import text file types into bundles (#107)
* import text file types into bundles

This is the first of what I think will be a few PRs for supporting non-js modules in workers. When esbuild scans the module graph, we collect references to some well known file types, mark them as external, and instead add them directly to the worker form upload. This prevents the need to configure this in wrangler.toml or whatever.

In the next PR, I'll do wasm support in a similar vein.

Once we have that in place, I'll probably work on a `--loader` flag to define custom types.

* Create pretty-monkeys-invent.md