enable eslint's `no-shadow` rule (#180)
* enable eslint's `no-shadow` rule

This PR enables eslint's `no-shadow` rule. I started a refactor of a couple of pieces and found myself getting confused with vars that claimed to be defined without it being so. The `no-shadow` rule prevents vars from being defined in inner scopes if they've already been defined in an outer scope.

This PR is mostly clean. It did lead to some wordy vars in `index.tsx`, but that's fine, it only made them more obvious.

I didn't bother doing this in `pages.tsx`, I'll leave that up to the pages team. Instead, I've just disabled the rule for that file.

* add a changeset

* stray commented out line

* Apply suggested changes from review