rename `--public` as `--experimental-public` (#214)
This PR is a breaking change. We'll keep it as a patch release tho.

We want to change the behaviour of `--public` to closer match Pages behaviour, and only really expose it in v2.1 (maybe even after that).

This PR renames the field as `experimental-public`, and adds warnings. In a future PR we'll actually change the behaviour.