feat: Workers Sites support for local mode (#336)
* chore: example-sites-app

An example app with a Workers Sites definition.

* feat: inline text-like files into the worker bundle

We were adding text-like modules (i.e. `.txt`, `.html` and `.pem` files) as separate modules in the Worker definition, but this only really 'works' with the ES module Worker format. This commit changes that to inline the text-like files into the Worker bundle directly.

We still have to do something similar with `.wasm` modules, but that requires a different fix, and we'll do so in a subsequent commit.

* feat: Sites support for local mode `wrangler dev`

This adds support for Workers Sites in local mode when running wrangler `dev`. Further, it fixes a bug where we were sending the `__STATIC_CONTENT_MANIFEST` definition as a separate module even with service worker format, and a bug where we weren't uploading the namespace binding when other kv namespaces weren't present.

* Update packages/wrangler/src/api/form_data.ts

Co-authored-by: Pete Bacon Darwin <pete@bacondarwin.com>

Co-authored-by: Pete Bacon Darwin <pete@bacondarwin.com>