refactor: tighten up the `dev` config typings
This internal change just makes the final `Config["dev"]` type expect to have non-optional properties, making it cleaner to use in code.
This is done, similarly to the `Environment` type by introducing `DevConfig` and partial `RawDevConfig` types,
which represent the runtime config and the `wrangler.toml` types respectively.
The `normalizeAndValidateDev()` helper now converts from the `RawDevConfig` to the `DevConfig`.