test: improve speed and resilience of pages upload tests (#1456)
The upload requests are run in parallel, so under some circumstances they can end up being completed in a different order.
The tests now are not affected by the order of the requests.

The page upload functionality uses `setTimeout()` with a 1 sec (or more) delay. This is now mocked to use `setImmediate()` to speed up these tests.