chore: don't kill `npm run dev` if any subtask fails (#1507)
We use `npm run dev` inside `packages/wrangler` to continuously build wrangler on any code change. This is useful when we're testing wrangler in any of the fixtures. However, it ends whenever we make a type error, which is common when experimenting, and interrupts the flow. It's better instead to log the error and carry on building as usual. This fix adds the flag to the concurrently task runner that enables this behaviour.