Add support for built-in `workerd` services (#435)
Specifically, this allows custom `network`, `external` and `disk`
services to be specified as service bindings. Requests to `network`
services are dispatched according to URL. Requests to `external`
services are dispatched to the specified remote server. Requests to
`disk` services are dispatched to an HTTP service backed by an
on-disk directory.

This PR also fixes a bug where custom function service bindings with
the same name in different Workers would dispatch all requests to the
first Workers' function.