Fix `--(experimental)-local` on Linux (#2275)
* Use different port for proxy and local runtime inspector servers

Previously, the Wrangler proxy and Node.js/workerd inspector servers
were listening on the same port. This somehow worked on macOS, but gives
a port already in use error on Linux.

Additionally, the Node.js inspector was configured to listen on
`0.0.0.0:<port>`. Aside from the security risk of opening an inspector
server on all interfaces, Node.js was logging `0.0.0.0:<port>` to the
console. Our RegExp to extract the debugger URL was looking for
`127.0.0.1:<port>` though, so we never actually connected to the
inspector.

* Bump `@miniflare/tre` to `3.0.0-next.8`

...fixing an issue where `setOptions()`/`dispose()` would not clean up
the previous workerd instance on Linux.

* Reduce verbosity of workerd startup failures