Allow custom `upstream`s to be specified (#568)
This adds back Miniflare 2's `upstream` option, which allows you to
customise the origin of incoming requests to the Worker. This is
useful when implementing proxy servers with Workers. Wrangler exposes
this option as the `--local-upstream` flag, so we'd like to support
this in Miniflare 3 to avoid regressing when we graduate.

Closes cloudflare/workers-sdk#2821
Closes DEVX-628