WC-1308 Reduce force-deletes unless explicitly confirmed (#3426)
Changes the default ?force param to the delete API call to be false.

Also uses the /scripts/{scriptName}/references and
/tails/by-consumer/{scriptName} endpoints to determine what resources
depend on the current script. We check for other scripts that use this
one as outbound worker (for dynamic dispatch), as a tail worker, as a
service binding, or use a durable object implemented by this script.
If any other resources rely on this worker, we require another step of
confirmation, describing the dependencies, and only then do we allow
doing a force delete.