rename `NPM_TOKEN` to `NPM_PUBLISH_TOKEN` (#3946)
We previously were using `NPM_TOKEN` because that's the name of the env var used in the .npmrc file that `changeset/action` produced for us. Now that we're using our own `.npmrc` file instead of letting it generate one for us, we're free to name it whatever. We already have a bunch of other prerelease workflows that expect `NPM_PUBLISH_TOKEN`, as well as it being the name of the github secret, so it's easiest to just change the name in our release step and `.npmrc` file to match.