fix: allow `wrangler pages dev` sessions to be reloaded (#4054)
#3951 introduced a change that closed the bootstrapper's IPC channel
when Wrangler exited. Whenever `wrangler (pages) dev` reloads, an
event is sent on this channel to indicate the dev server is ready.
Unfortunately, `wrangler pages dev`'s handler doesn't wait for the
dev session to exit before resolving its returned promise, meaning
the channel is closed as soon as the dev server is ready, not when
Wrangler exits. This means that ready events sent from reloads result
in an `ERR_IPC_CHANNEL_CLOSED` that crashes Wrangler. :(