[wrangler] Clean up compatibility fallback warnings (#4209)
* Only show compatibility date fallback warning if new version available

When a compatibility date greater than the installed `workerd` version
is selected, Miniflare logs a warning. The intention here is to warn
users if they've set a compatibility date, but haven't updated
`workerd`, so features they've requested won't be available.
The problem is that this warning is only actionable if there's an
update available. This change suppresses the warning if there's no
new version available.

* Use installed `workerd` version as default compat date instead of now

`wrangler dev` would previously use the current date as the default
compatibility date if one wasn't set in `wrangler.toml`. Before the
previous commit, this would show a warning if it was greater than the
installed `workerd` version. Whilst using the current date as the
default would no longer log a warning with this change, if there was
an update it would. This change sets the default to the `workerd`
version instead.