fix: return 4xx error from playground preview worker on invalid tokens (#4904)
Previously, we were passing user-controllable token inputs
(`X-CF-Token` header, `user` search param, `token` cookie) directly to
the `DurableObjectNamespace#idFromString()` function. This validated
the token to be a valid Durable Object ID, but threw a reportable
`TypeError`, generating a 500 response and increasing the errored
request count. This is a user error, so shouldn't be reported. This
change wraps calls to `idFromString()` inside `try`/`catch`es, and
re-throws non-reportable user errors that generate 4xx responses.
We could just validate that the tokens were 64 hex digits before
calling `idFromString()`, but this approach ensures the IDs are for
the `UserSession` object by validating the signature encoded in the
ID too.