Better messaging in remote mode (#5107)
* Partial refactor: remove zoneID from `getZoneIdHostAndRoutes()`

`zoneId` is only fetched in remote mode. Whether or not Wrangler is in remote mode can be decided further down the call stack than this function, and so when Wrangler is switched between local mode and remote mode the `zoneId` value returned from this function is stale. The `host` and `routes` returned from this function are always the same regardless of remote mode/local mode

* Remove `zone` and `zoneId` from props & typing

* Get zone in `remote.tsx` and only switch host when running a zone preview

* More detailed error messages + tests

* Fix snapshots

* changesets

* Remove .only

* fix build

* Make e2e validation stricter

* Fix tests & push inferred host down in call stack

* setTimeout -> retry

* strip ansi