test: fix vitest-pool-workers tests that rely on local workspace packages (#6057)
These tests create a fake package that relies upon the local versions of vitest-pool-workers and its dependencies. To make this work we package up all the local dependencies and add them to pnpm overrides.

Unfortunately we missed the transitive `@cloudflare/kv-asset-handler` dependency, so when that got a version bump the tests broke.

This is the simple, naive fix - adding in this package to those being packed. We ought to find a better solution that doesn't require manual dependency tracking.