fix: make `wrangler deploy` not call Workflows API if it binds to a different script (#7964)
Basically, if a worker bounds to a Workflow defined in another script,
it would do a PUT call to the Workflow API, overriding the config and
because it would change the `script_name`, it would break new instances
(because this one will probably not have the class exported).

This commit makes it so that Wrangler does not call the API if the
`script_name` is different.