Refactor pipelines commands into separate files, rename some options (#7769)
* Refactor pipelines commands into separate files, rename some options

**Pipelines is currently in closed beta, renaming these without
providing aliases _should_ have no impact**

Moving each of the subcommand handlers and options for each subcommand into its own
file. This also renames many of the parameters to be more specific.

The following parameters have been renamed:

| Previous Name | New Name |
| ---- | ---- |
| access-key-id | r2-access-key-id |
| secret-access-key | r2-secret-access-key |
| transform | transform-worker |
| r2 | r2-bucket |
| prefix | r2-prefix |
| binding | enable-worker-binding |
| http | enable-http |
| authentication | require-http-auth |
| filename | file-template |
| filepath | partition-template |

Adds the following new option for `create` and `update` commands:

```
--cors-origins           CORS origin allowlist for HTTP endpoint (use * for any origin)  [array]
```

Closes https://jira.cfdata.org/browse/PIPE-160.

* Fix non-exported type

* Add snapshot for create pipeline

* Add support for R2 staging, print banner for list cmd

* Refactor our verifyBucketAccess function

This also replaces the use of retryOnError with a for-loop as
retryOnError wasn't working as expected.

* Update CORS Origins when provided

Previously this was nested under the `enableHttp` block which doesn't
make sense if HTTP is already enabled and the user just wants to update
the authentication or CORS settings.

* Add test for pipeline cors update

* Fix default for partition template

* remove defaults and rely on API defaults instead

* Use proper capitalization for Workers and Pipelines

Also split the previous changeset into two.

* Apply suggestions from code review

Co-authored-by: Edmund Hung <edmund@cloudflare.com>

* Drop metrics calls, use chalk for CLI flag groups

Based on PR feedback, drop the metrics calls and use chalk.bold for CLI
flag groups. Snapshots were updated for tests.

* Update .changeset/short-kids-hunt.md

Co-authored-by: Edmund Hung <edmund@cloudflare.com>

---------

Co-authored-by: Edmund Hung <edmund@cloudflare.com>
Co-authored-by: emily-shen <69125074+emily-shen@users.noreply.github.com>